apiapp/requirements.txt
Django>=5.0.1
djangorestframework>=3.14.0
dj-rest-auth[with_social]>=5.0.2
django-cors-headers>=4.3.1
Pillow>=10.2.0
